backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
python-dotenv
fastapi-cors
passlib[bcrypt]
python-jose[cryptography]